rename reducers to store
